vm refactor - BlockchainUpdate
